+++ CHANGES.md
@@ -13,6 +13,7 @@ and this project adheres to
 - Add results command to list all test results
 - Add log command to emit log entries from .gob files
 - Add rm command to remove results and update latest symlink
+- Add buffering support for file data (see bufmem/buftmp in node.System.Stdout)
 
 ### Changed